Moved login information to the main Controller class.
