+++ src/controller/Controller.java
@@ -1,10 +1,10 @@
 package controller;
 
-import gui.EventGUI;
-import gui.Login;
-import gui.MailDelivery;
-import gui.TransportDiscontinued;
+import gui.*;
 import javafx.stage.Stage;
+import model.User;
+
+import java.util.List;
 
 public class Controller {
 	public static final String MAIL = "MAIL";
@@ -16,6 +16,7 @@ public class Controller {
 
 	// Global System Components
 	private UserDatabase userDatabase;
+	private User loggedIn;
 
 	public Controller(Stage primaryStage) {
 		this.primaryStage = primaryStage;
@@ -44,5 +45,38 @@ public class Controller {
 
 	public UserDatabase getUserDatabase() {
 		return userDatabase;
+	}
+
+	/**
+	 * @return the {@link User} logged in.
+	 */
+	public User getLoggedInUser() {
+		return loggedIn;
+	}
+
+	/**
+	 * Set the current logged in user. To retrieve a user to set for this method, use the
+	 * {@link UserDatabase#getUser(String)} or implicitly through {@link UserDatabase#getUsers()}
+	 *
+	 * @param loggedIn set the user that is logged in to the application
+	 */
+	public void setLoggedInUser(User loggedIn) {
+		this.loggedIn = loggedIn;
+	}
+
+	public void login(String inputUsername, String inputPassword) {
+		List<User> users = userDatabase.getUsers();
+		for(User user : users) {
+			if(user.getUsername().equals(inputUsername)
+					&& user.getPassword().equals(inputPassword)) {
+				setLoggedInUser(user);
+				handleEvent(Controller.EVENTGUI);
+				return;
+			}
+		}
+
+		// Display Error message here if user
+		AlertBox.display("Incorrect Login Information",
+				"Please input the correct credentials to login to KPSmart");
 	}
 }
