+++ src/gui/Login.java
@@ -65,26 +65,10 @@ public class Login implements EventHandler<ActionEvent>{
 	@Override
 	public void handle(ActionEvent event) {
 		if(event.getSource() == loginButton){
-			//System.out.println("LoginButton");
-			//System.out.println(usernameText.getText());
-			//System.out.println(passwordText.getText());
-			String input_Username = usernameText.getText();
-			String input_password = passwordText.getText();
+			String inputUsername = usernameText.getText();
+			String inputPassword = passwordText.getText();
 
-			List<User> users = controller.getUserDatabase().getUsers();
-			for(User user : users) {
-				if(user.getUsername().equals(input_Username)
-				&& user.getPassword().equals(input_password)) {
-				    controller.getUserDatabase().setLoggedInUser(user);
-					controller.handleEvent(Controller.EVENTGUI);
-					return;
-				}
-			}
-
-			// Display Error message here
-            AlertBox.display("Incorrect Login Information",
-                    "Please input the correct credentials to login to KPSmart");
-
+			controller.login(inputUsername, inputPassword);
 		}
 
 	}
